Revert "Revert "   - Remove all in-memory repository implementations and test files""
This reverts commit 34bef742fc40d798ad4dfe5783a087ade6d7b5c3.

+++ internal/repository/transaction/repository.go
@@ -14,7 +14,7 @@ type Repository interface {
 	// tx must be a live transaction when the insert should be part of a larger
 	// atomic operation; pass nil to run outside any transaction.
 	// amount and txType are pre-computed by the service layer.
-	Create(ctx context.Context, tx *sql.Tx, accountID, operationTypeID int64, amount float64, txType string) (*model.Transaction, error)
+	Create(ctx context.Context, tx *sql.Tx, accountID, operationTypeID int64, amount int64, txType string) (*model.Transaction, error)
 
 	// FindByID retrieves a transaction by its primary key.
 	FindByID(ctx context.Context, transactionID int64) (*model.Transaction, error)
